requirements.txt
requests~=2.27.1
discord.py~=1.7.3
tabulate~=0.8.9
pathos~=0.2.8